requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
boto3==1.29.7
python-dotenv==1.0.0
openai==2.8.1
kafka-python==2.0.2
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
requests==2.31.0
